requirements.txt
wasabi~=0.8.2
spleeter
dearpygui~=0.6.415
sox~=1.4.1
numpy~=1.18.5
tqdm~=4.61.1
omegaconf~=2.1.0
pyyaml~=5.4.1
future~=0.18.2